wip: JSW format to Tiled format

+++ rust/libs/open_jsw_tiled/src/tiled/point.rs
@@ -1,12 +1,12 @@
-use nanoserde::DeJson;
+use nanoserde::{DeJson, SerJson};
 
 /// Represents a Point in Tiled.
-#[derive(Clone, Debug, Default, DeJson)]
+#[derive(Clone, Debug, Default, DeJson, SerJson)]
 #[nserde(default)]
 pub struct Point {
     /// X coordinate in pixels.
-    pub x: f64,
+    pub x: i32,
 
     /// Y coordinate in pixels.
-    pub y: f64,
+    pub y: i32,
 }
